Graph カスタムフック切り出し
